requirements.txt
numpy~=2.2.0
soundfile~=0.12.1
matplotlib~=3.10.0
transformers~=4.47.0
yaml~=0.2.5
pyyaml~=6.0.2
huggingface-hub~=0.26.5
scipy~=1.14.1
requests~=2.32.3
tqdm~=4.67.1
future~=1.0.0
zlib~=1.2.13
referencing~=0.35.1
pygments~=2.18.0
bleach~=6.2.0
beautifulsoup4~=4.12.3
soupsieve~=2.6
h11~=0.14.0
defusedxml~=0.7.1
ipython~=8.30.0
pillow~=11.0.0
jupyterlab-pygments~=0.3.0
pip~=24.2
attrs~=24.2.0
wheel~=0.44.0
pysocks~=1.7.1
openssl~=3.0.15
tornado~=6.4.2
jinja2~=3.1.4
pyzmq~=26.2.0
colorama~=0.4.6
cffi~=1.17.1
setuptools~=75.1.0
comm~=0.2.2
traitlets~=5.14.3
fqdn~=1.5.1
parso~=0.8.4
jedi~=0.19.2
notebook~=7.3.1
idna~=3.10
typing_extensions~=4.11.0
sniffio~=1.3.1
arrow~=1.3.0
python-dateutil~=2.9.0post0
babel~=2.16.0
pytz~=2024.2
httpcore~=1.0.7
httpx~=0.28.1
pywin32~=308
json5~=0.10.0
llvmlite~=0.43.0
numba~=0.60.0
psutil~=6.1.0
ipykernel~=6.29.5
nbformat~=5.10.4
overrides~=7.7.0
packaging~=24.2
platformdirs~=4.3.6
regex~=2024.11.6
mpmath~=1.3.0
sympy~=1.13.1
python~=3.12.7
decorator~=5.1.1
urllib3~=2.2.3
argon2-cffi-bindings~=21.2.0
webencodings~=0.5.1
tinycss2~=1.4.0
markupsafe~=2.1.3
winpty~=0.4.3
pywinpty~=2.0.14
certifi~=2024.8.30
debugpy~=1.8.10
pandas~=2.2.3
ffmpeg-python~=0.2.0
six~=1.17.0
stack_data~=0.2.0
stack-data~=0.6.3
matplotlib-inline~=0.1.7
jupyter_client~=8.6.0
jupyter-client~=8.6.3
prompt_toolkit~=3.0.43
prompt-toolkit~=3.0.48
wcwidth~=0.2.13
anyio~=4.7.0
jupyter_core~=5.7.2
nbclient~=0.10.1
fastjsonschema~=2.21.1
jsonschema~=4.23.0
jupyter_server~=2.14.1
jupyter-server~=2.14.2
jupyterlab~=4.3.3
jupyterlab_server~=2.27.3
notebook-shim~=0.2.4
charset-normalizer~=3.4.0
asttokens~=3.0.0
contourpy~=1.3.1
executing~=2.1.0
fonttools~=4.55.3
nbconvert~=7.16.4
pandocfilters~=1.5.1
mistune~=3.0.2
jupyterlab_pygments~=0.1.2
pure_eval~=0.2.2
pure-eval~=0.2.3
pycparser~=2.22
pyparsing~=3.2.0
tokenizers~=0.21.0
websocket-client~=1.8.0
terminado~=0.18.1
webcolors~=24.11.1
jsonpointer~=3.0.0
uri-template~=1.3.0
isoduration~=20.11.0
rfc3339-validator~=0.1.4
jsonschema-specifications~=2024.10.1
async-lru~=2.0.4
cycler~=0.12.1
kiwisolver~=1.4.7
send2trash~=1.8.3
more-itertools~=10.5.0
jupyter-lsp~=2.2.5
safetensors~=0.4.5
jupyter_events~=0.10.0
python-json-logger~=2.0.7
prometheus_client~=0.21.0
prometheus-client~=0.21.1
jupyter_server_terminals~=0.4.4
jupyter-server-terminals~=0.5.3
ml-collections~=1.0.0
absl-py~=2.1.0
polars~=1.18.0
win_inet_pton~=1.1.0
torchinfo~=1.8.0
whisper~=1.1.10
openai-whisper~=20240930
pydub~=0.25.1